% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_genes}
\alias{all_genes}
\title{get all known gene identifiers from the selected assembly - the universe}
\usage{
all_genes(assembly = NULL)
}
\arguments{
\item{assembly}{selected assembly name}
}
\description{
get all known gene identifiers from the selected assembly - the universe
}
