% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_assemblies}
\alias{available_assemblies}
\title{return a dataframe of package names, assemblies and species}
\usage{
available_assemblies()
}
\description{
return a dataframe of package names, assemblies and species
}
