% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lichengo_packages}
\alias{lichengo_packages}
\title{List all packages in the lichengo collection}
\usage{
lichengo_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include lichengo in the list?}
}
\description{
List all packages in the lichengo collection
}
\examples{
tidyverse_packages()
}
