% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mapping}
\alias{mapping}
\title{get mapping between genes and GO terms}
\usage{
mapping(assembly = NULL)
}
\arguments{
\item{assembly}{selected assembly name}
}
\description{
get mapping between genes and GO terms
}
