% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_terms}
\alias{read_terms}
\title{get all known gene identifiers from the selected assembly and the annotations}
\usage{
read_terms(assembly = NULL)
}
\arguments{
\item{assembly}{selected assembly name}
}
\description{
get all known gene identifiers from the selected assembly and the annotations
}
